// angular-guusto/src/app/shared/service/crud/crud-service-resource.enum.ts
export enum CrudServiceResource {

    PERSIST = "CrudService.Persist",
    UPDATE = "CrudService.Update",
    REMOVE = "CrudService.Remove",
    REMOVE_BY_ID = "CrudService.RemoveById",
    FIND = "CrudService.Find",
    CONTAINS = "CrudService.Contains",
    FIND_ALL = "CrudService.FindAll",
    FIND_BY_ID = "CrudService.FindById",
    COUNT = "CrudService.Count",
    SEARCH = "CrudService.Search"

}